requirements.txt
transformers
torch
accelerate
bitsandbytes
outlines
# openai
PyYAML
tqdm
sentence-transformers  # if you want semantic matching

